lambdr: Create a Runtime for Serving Containerised R Functions on AWS Lambda

This package provides an R runtime for the AWS Lambda serverless compute service. It is intended to be used to create containers that can run on AWS Lambda. lambdr provides the necessary functionality for handling the various endpoints required for accepting new input and sending responses.

This package is unofficial. Its creators are not affiliated with Amazon Web Services, nor is its content endorsed by Amazon Web Services. Lambda, API Gateway, EventBridge, CloudWatch, and SNS are services of Amazon Web Services.

To see an example of how to use this package to create a runtime, refer to vignette("lambda-runtime-in-container", package = "lambdr").

The following invocation types have been implemented to some degree. Refer to the vignettes or the package website for more information.

Alternatively, custom functions can be provided for parsing event content and serialising results. Refer to lambda_config for more information.

Direct invocations

REST API Gateway invocations

vignette("api-gateway-invocations", package = "lambdr")

HTML API Gateway invocations

vignette("api-gateway-invocations", package = "lambdr")

EventBridge invocations

vignette("eventbridge-and-sns-invocations", package = "lambdr")

SNS invocations

vignette("eventbridge-and-sns-invocations", package = "lambdr")