Post an event with a result to the response endpoint

This function will first serialise the event result according to its class by dispatching through serialise_result.The result of that serialisation is posted as is to the response endpoint; this function will not perform any JSON serialisation, for example.

post_result(event, config)

Arguments

event

the response received from querying the next invocation endpoint.

config

A list of configuration values as created by the lambda_config function.

Invocations via an API Gateway

Events coming from an API Gateway need to be treated a little differently, both in parsing the event content and in posting the results. Refer to vignette("api-gateway-invocations", package = "lambdr") for details.