Raise an error when there's a missing request ID

A missing Request ID is an unusual error. It occurs after the runtime has been initialised, and is technically an invocation error. However it's impossible to post the error to the invocation error endpoint as that requires a Request ID.

We define a separate condition for it here, such that instead of stop we can use stop_missing_request_id. This allows us to catch and treat the error in a unique way. The handle_missing_request_id function accepts an error caught by tryCatch, logs it, and then returns a standard error with a sensible message.

stop_missing_request_id(message)

handle_missing_request_id(e)

Arguments

message
character message, such as an error description, to include with the condition.

e
error caught by tryCatch